+++ lib/index.js
@@ -1,30 +1,28 @@
 var _ = require('lodash');
 
-module.exports = function () {
+module.exports = function (soapFlag) {
 
-  var threadneedle = {
+    var threadneedle = {
 
-    // The key method
-    addMethod: require('./addMethod'),
+        // The key method
+        addMethod: require('./addMethod')(soapFlag),
 
-    // Global hook to set global options below
-    global: require('./global'),
+        // Global hook to set global options below
+        global: require('./global'),
 
-    // Default global settings (can be overridden)
-    _globalOptions: {
+        // Default global settings (can be overridden)
+        _globalOptions: {
 
-      // TODO Good requests generall come back with a '2xx'
-      // expects: {
-      //   statusCode: 200
-      // }
+            // TODO Good requests generall come back with a '2xx'
+            // expects: {
+            //   statusCode: 200
+            // }
 
-    }
-    
-  };
+        }
+
+    };
 
 
-  return threadneedle;
+    return threadneedle;
 
 };
-
-
