+++ package.json
@@ -30,6 +30,7 @@
     "mout": "^0.11.1",
     "mustache": "^2.2.1",
     "needle": "^1.4.2",
+    "soap": "^0.21.0",
     "when": "^3.7.7",
     "winston": "^2.1.1"
   },
